requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
deepdiff>=6.7.1
jsonschema>=4.21.1

# Cloud storage
supabase>=2.28.0

# CLI output
tabulate>=0.9.0

# Optional: Audio transcription (uncomment if needed)
# openai-whisper>=20231117
